Move NETWORK_DATA out of the UI
