+++ packages/types/src/fuse.ts
@@ -1,0 +1,73 @@
+import { BigNumber } from "ethers";
+
+export interface FuseAsset {
+  cToken: string;
+  plugin?: string;
+
+  borrowBalance: BigNumber;
+  supplyBalance: BigNumber;
+  liquidity: BigNumber;
+
+  membership: boolean;
+
+  underlyingName: string;
+  underlyingSymbol: string;
+  underlyingToken: string;
+  underlyingDecimals: BigNumber;
+  underlyingPrice: BigNumber;
+  underlyingBalance: BigNumber;
+
+  collateralFactor: BigNumber;
+  reserveFactor: BigNumber;
+
+  adminFee: BigNumber;
+  fuseFee: BigNumber;
+
+  borrowRatePerBlock: BigNumber;
+  supplyRatePerBlock: BigNumber;
+
+  totalBorrow: BigNumber;
+  totalSupply: BigNumber;
+
+  isBorrowPaused: boolean;
+  isSupplyPaused: boolean;
+}
+
+export interface NativePricedFuseAsset extends FuseAsset {
+  supplyBalanceNative: number;
+  borrowBalanceNative: number;
+
+  totalSupplyNative: number;
+  totalBorrowNative: number;
+
+  liquidityNative: number;
+  utilization: number;
+}
+
+export interface FusePoolData {
+  id: number;
+  assets: NativePricedFuseAsset[];
+  creator: string;
+  comptroller: string;
+  name: string;
+  totalLiquidityNative: number;
+  totalAvailableLiquidityNative: number;
+  totalSuppliedNative: number;
+  totalBorrowedNative: number;
+  totalSupplyBalanceNative: number;
+  totalBorrowBalanceNative: number;
+  blockPosted: BigNumber;
+  timestampPosted: BigNumber;
+  underlyingTokens: string[];
+  underlyingSymbols: string[];
+  whitelistedAdmin: boolean;
+  utilization: number;
+}
+
+export interface FusePool {
+  name: string;
+  creator: string;
+  comptroller: string;
+  blockPosted: number;
+  timestampPosted: number;
+}
