+++ packages/ui/components/shared/SwitchNetworkButton.tsx
@@ -1,21 +1,21 @@
 import { Button, Center, Img, Text, useDisclosure } from '@chakra-ui/react';
+import { ChainConfig } from '@midas-capital/types';
 import React, { LegacyRef, useEffect, useState } from 'react';
 
 import SwitchNetworkModal from '@ui/components/shared/SwitchNetworkModal';
 import { useRari } from '@ui/context/RariContext';
 import { useIsSmallScreen } from '@ui/hooks/useScreenSize';
-import { getChainMetadata } from '@ui/networkData/index';
-import { ChainMetadata } from '@ui/types/ChainMetaData';
+import { getChainConfig } from '@ui/networkData/index';
 
 const SwitchNetworkButton: React.FC = () => {
-  const [chainMetadata, setChainMetadata] = useState<ChainMetadata | undefined>();
+  const [chainMetadata, setChainConfig] = useState<ChainConfig | undefined>();
 
   const { isOpen, onOpen, onClose } = useDisclosure();
   const isMobile = useIsSmallScreen();
   const { networkBtnElement, currentChain } = useRari();
 
   useEffect(() => {
-    setChainMetadata(getChainMetadata(currentChain.id));
+    setChainConfig(getChainConfig(currentChain.id));
   }, [currentChain]);
 
   return (
@@ -28,10 +28,16 @@ const SwitchNetworkButton: React.FC = () => {
       <Center>
         {chainMetadata && (
           <>
-            {chainMetadata.img && (
-              <Img width="25px" height="25px" borderRadius="50%" src={chainMetadata.img} alt="" />
+            {chainMetadata.specificParams.metadata.img && (
+              <Img
+                width="25px"
+                height="25px"
+                borderRadius="50%"
+                src={chainMetadata.specificParams.metadata.img}
+                alt=""
+              />
             )}
-            {!isMobile && <Text ml={2}>{chainMetadata.name}</Text>}
+            {!isMobile && <Text ml={2}>{chainMetadata.specificParams.metadata.name}</Text>}
           </>
         )}
       </Center>
