+++ packages/ui/utils/web3Providers.ts
@@ -1,21 +1,10 @@
 import { JsonRpcProvider, Web3Provider } from '@ethersproject/providers';
-import { bsc, chapel, ganache, moonbeam, neondevnet, polygon } from '@midas-capital/chains';
 import { MidasSdk } from '@midas-capital/sdk';
-import { ChainConfig } from '@midas-capital/types';
 
-import { NETWORK_DATA } from '@ui/networkData/index';
-
-const chainIdToConfig: { [chainId: number]: ChainConfig } = {
-  [bsc.chainId]: bsc,
-  [polygon.chainId]: polygon,
-  [moonbeam.chainId]: moonbeam,
-  [neondevnet.chainId]: neondevnet,
-  [chapel.chainId]: chapel,
-  [ganache.chainId]: ganache,
-};
+import { chainIdToConfig } from '@ui/types/ChainMetaData';
 
 export function providerURLForChain(chainId: number) {
-  const network = Object.values(NETWORK_DATA).find((each) => each.chainId === chainId);
+  const network = chainIdToConfig[chainId].specificParams.metadata;
   if (!network) {
     throw new Error(`Unable to get providerUrlForChain() for chainId: ${chainId}`);
   }
